Backend/requirements.txt
torch
torchvision
opencv-python
pillow
fastapi
uvicorn
python-multipart
